[FIX] website_sale: disable link from pivot view for sale.report
Before this commit, linking was enable for pivot view of model sale.report.
User that click on a cell will be redirected on a default view list/form
with all fields.

Now linking is disabled.

If you want to open the sale order on click, you need to start with the
pivot view from model sale.order.

This commit closes #23445