[FIX] account: make Journal Item company read-only
PR #16826 (rev a91fedfe8a624713bb76f04ba28d66ea9b262f30) added the
`company_id` field in the Journal Entry form.

Unfortunately that field was not marked read-only, so creating a move
would write to this field, triggering a useless and possibly very long
recomputation of many related fields based on the company of the
account. It must be read-only, its value solely driven by the value of
the account_id field.